make programmatic links in svgs use react router (#191)
Use react router to programmatically change url on click of heatmap and volcano plot elements (which can't use the typical react router <Link> components), instead of setting window.location. This makes it so the whole app doesn't have to reload when clicking on these svg links, or when going backward after doing so.